% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_risk_development.R
\name{aggregate_risk_development}
\alias{aggregate_risk_development}
\title{Create an Interactive Risk Share Bar-plot (Upside or Downside)}
\usage{
aggregate_risk_development(
  file_paths_named_list,
  infl_col = c(16),
  upside_col = c(17:22),
  downside_col = c(24:29),
  xlab = "Survey Date",
  ylab = "Share of Total Risk",
  title = "Share of Aggregated Upside vs Downside Risk Over Time"
)
}
\arguments{
\item{infl_col}{A vector stating in which column of the file the inflation data are.}

\item{upside_col}{A vector stating in which columns of the file the upside risk data are.}

\item{downside_col}{A vector stating in which columns of the file the downside risk data are.}

\item{xlab}{A character string specifying the x-axis label (optional).}

\item{ylab}{A character string specifying the y-axis label (optional).}

\item{title}{A character string specifying the title of the graph (optional).}

\item{df}{A data frame containing survey data.}
}
\value{
A \code{plotly} interactive bar plot.
}
\description{
This function reads SMA survey data from multiple files and visualizes the composition of
upside or downside risk over time as an interactive stacked bar plot.
}
\examples{
\dontrun{
files <- prepare_file_list(c("May 25", "Jun 25"))
aggregate_risk_development(files)
}

}
\author{
Ole Paech
}
