% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk_Factors_Composite.R
\name{aggregate_risk_factors}
\alias{aggregate_risk_factors}
\title{Create Simplified Interactive Risk Factors Bar Plot}
\usage{
aggregate_risk_factors(
  df,
  infl_col = c(16),
  upside_col = c(17:21),
  downside_col = c(23:27),
  xlab = "",
  ylab = "Average Inflation Expectation",
  title = ""
)
}
\arguments{
\item{df}{A data frame containing survey data.}

\item{infl_col}{A vector stating in which column of the file the inflation data are.}

\item{upside_col}{A vector stating in which columns of the file the upside risk data are.}

\item{downside_col}{A vector stating in which columns of the file the downside risk data are.}

\item{xlab}{A character string specifying the x-axis label (optional).}

\item{ylab}{A character string specifying the y-axis label (optional).}

\item{title}{A character string specifying the title of the graph (optional).}
}
\value{
A \code{plotly} interactive bar plot.
}
\description{
This function creates an interactive bar plot showing the combined upside and downside risks
alongside the average inflation expectation, without breaking them into individual components.
}
\examples{
\dontrun{
path <- load_participant_files()
data <- readxl::read_excel(path)
aggregate_risk_factors(data)
}

}
\author{
Ole Paech
}
