% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_line.R
\name{bar_line}
\alias{bar_line}
\title{Create a bar and line plot showing median and IQR over time}
\usage{
bar_line(data)
}
\arguments{
\item{data}{A data.frame with the survey data.}
}
\value{
A Plotly object with dual-axis bar and line chart.
}
\description{
Create a bar and line plot showing median and IQR over time
}
\examples{
path <- load_participant_files(subfolder = "Current")
data <- readxl::read_excel(path)
bar_line(data)

}
\author{
Ole Paech
}
