% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dotplot_all_categories.R
\name{dotplot}
\alias{dotplot}
\title{Create a jittered dotplot of rates by month and category}
\usage{
dotplot(data, category)
}
\arguments{
\item{data}{A data frame containing survey data.}

\item{category}{A string indicating the grouping category. Must be one of
\code{"Profession"}, \code{"Experience"}, or \code{"Nationality"}.}
}
\value{
A Plotly interactive plot object.
}
\description{
This function takes NBS survey data and creates a jittered dotplot
showing individual rate observations by month, colored by a
respondent category (profession, experience, or nationality)
}
\examples{
\dontrun{
  path <- load_participant_files()
  data <- readxl::read_excel(path)
  dotplot(data, category = "Experience")
}

}
\author{
Ole Paech
}
