% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_current_dfr.R
\name{get_current_dfr}
\alias{get_current_dfr}
\title{Scrape the current ECB Deposit Facility Rate (DFR)}
\usage{
get_current_dfr()
}
\value{
A numeric value representing the current DFR (in percent).
}
\description{
Downloads and parses the current ECB Deposit Facility Rate from
TradingEconomics.com using a fixed XPath.
}
\examples{
\dontrun{
get_current_dfr()
}


}
\author{
Ole Paech
}
