% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_bar_category.R
\name{group_bar_categroy}
\alias{group_bar_categroy}
\title{Grouped Bar Plot by Category with Median, Mean and Mode}
\usage{
group_bar_categroy(data, category)
}
\arguments{
\item{data}{A data frame containing the survey data.}

\item{category}{A character string specifying the category to group by.
Must be one of \code{"Profession"}, \code{"Experience"}, or \code{"Nationality"}.}
}
\value{
A Plotly bar plot object.
}
\description{
This function takes NBS survey data and creates a grouped bar plot
showing Median, Mean and Mode of survey values by a respondent category
(profession, experience, or nationality), colored by month.
}
\examples{
\dontrun{
  path <- load_participant_files()
  data <- readxl::read_excel(path)
  group_bar_categroy(data, category = "Profession")
}

}
\author{
Ole Paech
}
