% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hover_boxplot.R
\name{hover_boxplot}
\alias{hover_boxplot}
\title{Interactive Boxplot for DFR expectations}
\usage{
hover_boxplot(data)
}
\arguments{
\item{data}{A data frame containing survey responses.}
}
\value{
A \code{plotly} interactive boxplot object.
}
\description{
Generates a Plotly boxplot with median and percentile statistics for NBS survey data.
}
\examples{
\dontrun{
path <- load_participant_files()
data <- readxl::read_excel(path)
hover_boxplot(data)
}

}
\author{
Ole Paech
}
