% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hover_violin.R
\name{hover_violin}
\alias{hover_violin}
\title{Interactive violin-plot for DFR expectations}
\usage{
hover_violin(data)
}
\arguments{
\item{data}{A data frame containing survey responses.}
}
\value{
A \code{plotly} interactive violin-plot object.
}
\description{
Generates a Plotly violin-plot with median and percentile statistics for NBS survey data.
}
\examples{
\dontrun{
path <- load_participant_files()
data <- readxl::read_excel(path)
hover_boxplot(data)
}

}
\author{
Ole Paech
}
