% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_risk_history.R
\name{inflation_risk_history}
\alias{inflation_risk_history}
\title{Visualize Historical Inflation Risk Perception}
\usage{
inflation_risk_history(files_with_labels)
}
\arguments{
\item{files_with_labels}{A chronological named list, where each element is the path
to an Excel file and the name is the corresponding survey date label.}
}
\value{
An interactive plotly object visualizing the inflation expectation
and risk factor contribution/s over time.
}
\description{
This function processes labeled Excel survey files and visualizes
the historical development of upside and downside inflation risk
in relation to inflation expectations using a stacked bar plot.
}
\examples{
\dontrun{
files <- prepare_file_list(c("May 25", "Jun 25"))
inflation_risk_history(files)
}

}
\author{
Ole Paech
}
