% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3d.R
\name{plot_3d}
\alias{plot_3d}
\title{3D Visualization of Median Inflation Expectations}
\usage{
plot_3d(data)
}
\arguments{
\item{data}{A data frame with variables including profession, experience, nationality,
and inflation expectation columns.}
}
\value{
A list of interactive plotly 3D scatter plots, one for each month.
}
\description{
This function creates interactive 3D scatter plots of median and mean inflation expectations
by profession, experience level, and nationality across different months.
}
\details{
The function is intended to process SMA survey data in wide format and create
one 3D plot per observed month.
}
\examples{
\dontrun{
path <- load_participant_files()
data <- readxl::read_excel(path)
plots <- plot_3d(data)
plots[[1]]  # Show first months plot
plots[[4]]  # Show Long-Run Plot
}

}
\author{
Ole Paech
}
