% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh_quarto_NBS.R
\name{refresh_quarto}
\alias{refresh_quarto}
\title{Render Quarto Website or Specific Indicator File}
\usage{
refresh_quarto(indicator = NULL, open = TRUE, background_job = FALSE)
}
\arguments{
\item{indicator}{Optional character string. Name of indicator file (without extension).}

\item{open}{Logical. Whether to open the rendered HTML in a browser. Defaults to TRUE.}

\item{background_job}{Logical. Run rendering as background job. Defaults to FALSE.}
}
\description{
Renders the full Quarto website or a single indicator \code{.qmd} file, depending on the input.
}
\author{
Martin Stefan (Deutsche Bundesbank), Ole Paech
}
