% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk_Factors.R
\name{risk_factors}
\alias{risk_factors}
\title{Create Interactive Risk Factors Bar Plot}
\usage{
risk_factors(df)
}
\arguments{
\item{df}{A data frame containing survey data. The 18th column should be inflation data,
columns 19-23 upside risk factors, and columns 24-28 downside risk factors.}
}
\value{
A \code{plotly} interactive bar plot.
}
\description{
This function takes a data frame from SMA survey results and creates an interactive bar plot
showing average and weighted upside and downside risk factors alongside inflation expectation.
}
\examples{
\dontrun{
path <- load_participant_files()
data <- readxl::read_excel(path)
risk_factors(data)
}

}
\author{
Ole Paech
}
