% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk_Development_Area.R
\name{risk_share_area}
\alias{risk_share_area}
\title{Plot Risk Share Area for Upside or Downside Risks}
\usage{
risk_share_area(file_paths_named_list, type = "Upside")
}
\arguments{
\item{file_paths_named_list}{Named list of file paths to Excel files.}

\item{type}{Character string specifying the risk type: "Upside" (default) or "Downside".}
}
\value{
A plotly interactive plot showing the share of total risk over time.
}
\description{
This function reads multiple Excel files with survey data and visualizes the composition of
upside or downside risk over time as an interactive stacked area plot.
}
\examples{
\dontrun{
file_paths <- prepare_file_list(c("May 25", "Jun 25"))
risk_share_area(file_paths, type = "Upside")    #Development of Upside Risks
risk_share_area(file_paths, type = "Downside")  #Development of Downside Risks
}

}
\author{
Ole Paech
}
