% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiderplot_category.R
\name{spiderplot_category}
\alias{spiderplot_category}
\title{Create a Radar Chart of Median Expectations by Category}
\usage{
spiderplot_category(
  data,
  category,
  Min = 1,
  Max = 4,
  rel_cols = c(10, 12, 14),
  show_title = FALSE
)
}
\arguments{
\item{data}{A data frame with participant survey responses.}

\item{category}{A character string specifying the category to group by.
Must be one of "Profession", "Experience", or "Nationality".}

\item{Min}{Numeric value for the minimum axis limit of the radar plot. Default is 1.}

\item{Max}{Numeric value for the maximum axis limit of the radar plot. Default is 4.}

\item{rel_cols}{A vector stating in which columns of the file the data to visualize are.}

\item{show_title}{A logical value indicating whether to display a title above the radar chart. Defaults to FALSE. Set to TRUE to add a title.}
}
\value{
A radar chart visualizing the median inflation expectations by group.
}
\description{
This function generates a spider (radar) plot of median expectations across months
for different groups based on a selected category (Profession, Experience, or Nationality).
}
\examples{
\dontrun{
path <- load_participant_files()
data <- readxl::read_excel(path)
spiderplot_category(data, category = "Profession")
}

}
\author{
Ole Paech
}
