% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_histogram.R
\name{grouped_histogram}
\alias{grouped_histogram}
\title{Grouped Histogram Plot of Expected Rates by Month}
\usage{
grouped_histogram(data)
}
\arguments{
\item{data}{Data frame containing the survey responses.}
}
\value{
A plotly interactive histogram plot.
}
\description{
Creates grouped histograms of expected rates over months with medians highlighted.
}
\examples{
\dontrun{
  path <- load_participant_files()
  data <- readxl::read_excel(path)
  grouped_histogram(data)
}

}
\author{
Ole Paech
}
