% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hover_bars.R
\name{hover_barplot}
\alias{hover_barplot}
\title{Create interactive bar chart of DFR expectations}
\usage{
hover_barplot(data)
}
\arguments{
\item{data}{A data frame containing survey responses.}
}
\value{
A \code{plotly} interactive barplot object.
}
\description{
Generates a Plotly bar chart with median, mode, and percentile statistics for NBS survey data.
}
\examples{
\dontrun{
path <- load_participant_files()
data <- readxl::read_excel(path)
hover_barplot(data)
}

}
\author{
Ole Paech
}
