% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_file_list.R
\name{prepare_file_list}
\alias{prepare_file_list}
\title{Prepare a named list of file paths}
\usage{
prepare_file_list(month_labels = c("May 25", "Jun 25"))
}
\arguments{
\item{month_labels}{A character vector of month labels to load (e.g., c("May 25", "Jun 25")).}
}
\value{
A named character vector of existing file paths, excluding any missing files.
}
\description{
Creates a single named list of file paths.
}
\examples{
\dontrun{
file_paths <- prepare_file_list(c("May 25", "Jun 25"))
}


}
\author{
Ole Paech
}
