% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk_Factors.R
\name{risk_factors}
\alias{risk_factors}
\title{Create Interactive Risk Factors Bar Plot}
\usage{
risk_factors(
  df,
  infl_col = c(16),
  upside_col = c(17:22),
  downside_col = c(24:29),
  xlab = "",
  ylab = "Average Inflation Expectation",
  title = ""
)
}
\arguments{
\item{df}{A data frame containing survey data.}

\item{infl_col}{A vector stating in which column of the file the inflation data are.}

\item{upside_col}{A vector stating in which columns of the file the upside risk data are.}

\item{downside_col}{A vector stating in which columns of the file the downside risk data are.}

\item{xlab}{A character string specifying the x-axis label (optional).}

\item{ylab}{A character string specifying the y-axis label (optional).}

\item{title}{A character string specifying the title of the graph (optional).}
}
\value{
A \code{plotly} interactive bar plot.
}
\description{
This function takes a data frame from SMA survey results and creates an interactive bar plot
showing average and weighted upside and downside risk factors alongside inflation expectation.
}
\examples{
\dontrun{
path <- load_participant_files()
data <- readxl::read_excel(path)
risk_factors(data)
}

}
\author{
Ole Paech
}
