% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Risk_Development_Bars.R
\name{risk_share_bars}
\alias{risk_share_bars}
\title{Create an Interactive Risk Share Bar-plot (Upside or Downside)}
\usage{
risk_share_bars(file_paths_named_list, type = "Upside")
}
\arguments{
\item{file_paths_named_list}{Named list of Excel file paths.}

\item{type}{Character string specifying risk type: either "Upside" (default) or "Downside".}
}
\value{
A \code{plotly} interactive bar plot showing risk share distribution over time.
}
\description{
This function reads SMA survey data from multiple files and visualizes the composition of
upside or downside risk over time as an interactive stacked bar plot.
}
\examples{
\dontrun{
file_paths <- prepare_file_list(c("May 25", "Jun 25"))
risk_share_bars(file_paths, type = "Upside")    #Development of Upside Risks
risk_share_bars(file_paths, type = "Downside")  #Development of Downside Risks
}

}
\author{
Ole Paech
}
