% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shaded Plot.R
\name{shaded_plot}
\alias{shaded_plot}
\title{Create an Interactive Shaded Plot of Survey Rates Over Time}
\usage{
shaded_plot(data)
}
\arguments{
\item{data}{A data frame with survey response data.}
}
\value{
A \code{plotly} interactive scatter plot with shaded interquartile range and median lines.
}
\description{
This function takes a data frame containing SMA survey responses with expected DFR rates over months,
cleans and reshapes the data, then creates an interactive plotly plot showing median values
along with the interquartile range as a shaded area.
}
\examples{
\dontrun{
path <- load_participant_files()
data <- readxl::read_excel(path)
shaded_plot(data)
}

}
\author{
Ole Paech
}
